TST: add tests for mixin
